Refactor list entries using child components
